Former-commit-id: f1c1453a0524101f6903d8b0248c9b3d191f04fe

+++ src/Core/BoundedContexts/CAPI/WB.Core.BoundedContexts.CAPI/Views/InterviewDetails/ValueQuestionViewModel.cs
@@ -23,7 +23,7 @@ namespace WB.Core.BoundedContexts.Capi.Views.InterviewDetails
 
         #region Overrides of QuestionViewModel
 
-        public override IQuestionnaireItemViewModel Clone(int[] propagationVector)
+        public override IQuestionnaireItemViewModel Clone(decimal[] propagationVector)
         {
             return new ValueQuestionViewModel(new InterviewItemId(this.PublicKey.Id, propagationVector),
                                                    this.SourceText, this.QuestionType, this.AnswerObject,
